backend/requirements.txt
# --- Web Framework & Server ---
fastapi==0.135.1
uvicorn==0.42.0
starlette==0.52.1

# --- AI & LLM (Gemini 1.5 Flash Support) ---
# Note: google-genai is the new SDK used in your GeminiService class
google-genai==1.68.0
google-generativeai==0.8.6

# --- Database & ORM ---
sqlalchemy==2.0.48
asyncpg==0.31.0

# --- Configuration & Env ---
python-dotenv==1.2.2
pydantic-settings==2.13.1

# --- Data Validation ---
pydantic==2.12.5
email-validator==2.3.0
annotated-types==0.7.0

# --- Security & Authentication ---
python-multipart==0.0.9
pyjwt==2.12.1
cryptography==46.0.5
google-auth==2.49.1
requests==2.32.5

# --- Utilities ---
httpx==0.28.1
httpcore==1.0.9
anyio==4.12.1
certifi==2026.2.25